Move util files from base folder to lib fodler.
